Revert "Explicit conversion and tests for new pbf_reader::data() function."
This reverts commit bd2ae468280ff46c946965035770c3c311c41699.

It contained more than it was supposed to contain.

+++ include/protozero/pbf_reader.hpp
@@ -251,7 +251,7 @@ public:
      * Get a view of the not yet read data.
      */
     data_view data() const noexcept {
-        return {m_data, static_cast<std::size_t>(m_end - m_data)};
+        return {m_data, m_end - m_data};
     }
 
     /**
